completed experminent 5.2

+++ Programs/section3.5/experiment5_2.c
@@ -4,12 +4,12 @@ int main(void)
 {
     float sum = 0;
 
-    for(int i = 1; i <=10; ++i)
+    for (int i = 1; i <= 10; ++i)
     {
-        sum += 1/i;
+        sum += (1 / (float)i);
     }
 
-    printf("sum is %.3f", sum);
+    printf("sum is %.7f", sum);
 
     return 0;
 }
